vi: extract types
