+++ src/core/hle/service/vi/vi_u.cpp
@@ -3,14 +3,15 @@
 
 #include "common/logging/log.h"
 #include "core/hle/service/vi/vi.h"
+#include "core/hle/service/vi/vi_types.h"
 #include "core/hle/service/vi/vi_u.h"
 
 namespace Service::VI {
 
 VI_U::VI_U(Core::System& system_, Nvnflinger::Nvnflinger& nv_flinger_,
            Nvnflinger::HosBinderDriverServer& hos_binder_driver_server_)
-    : ServiceFramework{system_, "vi:u"}, nv_flinger{nv_flinger_}, hos_binder_driver_server{
-                                                                      hos_binder_driver_server_} {
+    : ServiceFramework{system_, "vi:u"}, nv_flinger{nv_flinger_},
+      hos_binder_driver_server{hos_binder_driver_server_} {
     static const FunctionInfo functions[] = {
         {0, &VI_U::GetDisplayService, "GetDisplayService"},
         {1, nullptr, "GetDisplayServiceWithProxyNameExchange"},
